refactor(accesscore): lift initialadmin out of internal/, drop facades
initialadmin is a cohesive Cell subsystem (Bootstrapper/Sweep/Cleaner/
Lifecycle/Hasher) with its own public Option surface. Promoting it from
cells/accesscore/internal/initialadmin to cells/accesscore/initialadmin
eliminates the need for thin re-export facades in accesscore and lets
cmd/corebundle + examples + external tests invoke initialadmin.With*
directly. Matches the pattern of cells/accesscore/slices/ already being
public.

Removed:
- cells/accesscore/lifecycle_option.go (LifecycleOption alias + facade)
- cells/accesscore/password_hasher.go (PasswordHasher/BcryptHasher alias)

ref: kernel/cell.LifecycleContributor auto-discovery now resolved end-to-end
